# docs/reference/audittrace/openapi.yaml
...
components:
  schemas:
...
    ConsolePromptGroupItem:
      description: 'One prompt-group row (summary shape, no versions), as returned

        by the list-groups API.'
      properties:
        category:
          title: Category
          type: string
        command:
          anyOf:
          - type: string
          - type: 'null'
          title: Command
        created_at_ms:
          title: Created At Ms
          type: integer
        deleted_at_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Deleted At Ms
        group_id:
          title: Group Id
          type: string
        metadata:
          additionalProperties: true
          title: Metadata
          type: object
        name:
          title: Name
          type: string
        oneliner:
          title: Oneliner
          type: string
        production_prompt_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Production Prompt Id
        updated_at_ms:
          title: Updated At Ms
          type: integer
      required:
      - group_id
      - name
      - category
      - oneliner
      - created_at_ms
      - updated_at_ms
      title: ConsolePromptGroupItem
      type: object
    ConsolePromptGroupListResponse:
      description: 'Response from ``GET /console/prompts`` — cursor-paginated,

        newest-first, summary shape (no versions per row).'
      properties:
        items:
          items:
            $ref: '#/components/schemas/ConsolePromptGroupItem'
          title: Items
          type: array
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      title: ConsolePromptGroupListResponse
      type: object
    ConsolePromptGroupUpsertRequest:
      description: 'Request body for ``POST /console/prompts`` — create/update the

        caller''s own prompt group (upsert by ``group_id``).'
      properties:
        category:
          anyOf:
          - maxLength: 128
            type: string
          - type: 'null'
          title: Category
        command:
          anyOf:
          - maxLength: 128
            type: string
          - type: 'null'
          title: Command
        group_id:
          maxLength: 255
          minLength: 1
          title: Group Id
          type: string
        metadata:
          additionalProperties: true
          title: Metadata
          type: object
        name:
          maxLength: 512
          minLength: 1
          title: Name
          type: string
        oneliner:
          anyOf:
          - maxLength: 4096
            type: string
          - type: 'null'
          title: Oneliner
      required:
      - group_id
      - name
      title: ConsolePromptGroupUpsertRequest
      type: object
    ConsolePromptGroupWithVersionsItem:
      description: 'The full group shape, INCLUDING its version history — returned

        by ``GET /console/prompts/{group_id}`` and

        ``PATCH /console/prompts/{group_id}/production``.'
      properties:
        category:
          title: Category
          type: string
        command:
          anyOf:
          - type: string
          - type: 'null'
          title: Command
        created_at_ms:
          title: Created At Ms
          type: integer
        deleted_at_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Deleted At Ms
        group_id:
          title: Group Id
          type: string
        metadata:
          additionalProperties: true
          title: Metadata
          type: object
        name:
          title: Name
          type: string
        oneliner:
          title: Oneliner
          type: string
        production_prompt_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Production Prompt Id
        updated_at_ms:
          title: Updated At Ms
          type: integer
        versions:
          items:
            $ref: '#/components/schemas/ConsolePromptVersionItem'
          title: Versions
          type: array
      required:
      - group_id
      - name
      - category
      - oneliner
      - created_at_ms
      - updated_at_ms
      title: ConsolePromptGroupWithVersionsItem
      type: object
    ConsolePromptSetProductionRequest:
      description: 'Request body for ``PATCH /console/prompts/{group_id}/production``

        — mark ``prompt_id`` (an existing version of the group) as the

        group''s production version.'
      properties:
        prompt_id:
          maxLength: 255
          minLength: 1
          title: Prompt Id
          type: string
      required:
      - prompt_id
      title: ConsolePromptSetProductionRequest
      type: object
    ConsolePromptVersionItem:
      description: One prompt-version row, as returned by the console-prompts API.
      properties:
        created_at_ms:
          title: Created At Ms
          type: integer
        group_id:
          title: Group Id
          type: string
        metadata:
          additionalProperties: true
          title: Metadata
          type: object
        prompt_id:
          title: Prompt Id
          type: string
        text:
          title: Text
          type: string
        type:
          title: Type
          type: string
        updated_at_ms:
          title: Updated At Ms
          type: integer
        version:
          title: Version
          type: integer
      required:
      - prompt_id
      - group_id
      - text
      - type
      - version
      - created_at_ms
      - updated_at_ms
      title: ConsolePromptVersionItem
      type: object
    ConsolePromptVersionUpsertRequest:
      description: 'Request body for ``POST /console/prompts/{group_id}/versions`` —

        create/update the caller''s own prompt version (upsert by

        ``prompt_id``) attached to ``group_id``.'
      properties:
        metadata:
          additionalProperties: true
          title: Metadata
          type: object
        prompt_id:
          maxLength: 255
          minLength: 1
          title: Prompt Id
          type: string
        text:
          minLength: 1
          title: Text
          type: string
        type:
          default: text
          enum:
          - text
          - chat
          title: Type
          type: string
      required:
      - prompt_id
      - text
      title: ConsolePromptVersionUpsertRequest
      type: object
...
    ConsoleToolFavoriteListResponse:
      description: 'Response from ``GET /console/tool-favorites`` — the caller''s

        ENTIRE favorites list (no pagination; bounded by

        ``MAX_TOOL_FAVORITES``), oldest-first.'
      properties:
        items:
          items:
            $ref: '#/components/schemas/ConsoleToolFavoriteItem'
          title: Items
          type: array
      title: ConsoleToolFavoriteListResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    audittrace-oauth2:
      flows:
        authorizationCode:
          authorizationUrl: /protocol/openid-connect/auth
          scopes:
            audittrace:admin: 'Operator-grade administration: /memory/upload, /memory/index, hard-delete, configuration introspection.
              Optional scope on user-facing clients; never granted by default.'
            audittrace:assessment:ingest: Record a security self-assessment as audit events (ADR-058 recursive self-audit);
              a write grant distinct from the broad :audit read scope
            audittrace:audit: Read interactions, sessions, and tool-call audit rows
            audittrace:context: Read context-builder output for a query
            audittrace:index: Trigger semantic-store reindex (legacy dev client)
            audittrace:query: Issue chat completions and inference queries
            audittrace:scan:retrigger: Force a re-scan of an object in the ingestion content-control pipeline (ADR-048 operator
              surface). Admin-grade; never granted by default.
            memory:acl:read-own: Read effective permissions, accessible/public/sole-owned resource ids from the console-ACL
              store (Sovereign Authorization Layer EPIC, WU-1 — READ PATH ONLY) — the RLS-isolated store backing LibreChat's
              AclEntry records. Every answer is scoped to the CALLER's own resolved principal set (self + public); group principals
              are disabled at the schema level. This scope grants no write path (see memory:acl:write) and no access to any
              other memory layer.
            memory:acl:write: 'Write ACL grants through the console-ACL routes (Sovereign Authorization Layer EPIC, WU-2c):
              grant, revoke, modify, bulk and expire-by-predicate on the CALLER''s own resources. The database (RLS, migration
              032) decides who may write — a non-owner is refused and the refusal is audited. Does not open the read routes
              (memory:acl:read-own) and grants no access to any other memory layer.'
            memory:agents:read-own: 'Read your own agents (and batch-get-by-ids) from the console-agents store (Agents domain,
              MongoDB-elimination EPIC) — the RLS-isolated store backing LibreChat''s Agent records. Own-agents-only: sharing/marketplace
              is out of scope. Grants no access to another user''s agents or to any other memory layer.'
            memory:agents:write: Create/update/delete your own agents in the console-agents store (Agents domain, MongoDB-elimination
              EPIC). Grants no access to another user's agents (RLS-enforced) and no access to any other memory layer.
            memory:chat_projects:read-own: Read your own first-class chat-projects from the console-chat-projects store (Chat-Projects
              domain, MongoDB-elimination EPIC) — the RLS-isolated store backing LibreChat's chat-project groupings of conversations.
              Grants no access to another user's chat-projects or to any other memory layer.
            memory:chat_projects:write: Create/update/delete your own chat-projects in the console-chat-projects store (Chat-Projects
              domain, MongoDB-elimination EPIC). Grants no access to another user's chat-projects (RLS-enforced) and no access
              to any other memory layer.
            memory:conversation_tags:read-own: 'Read your own conversation-tags from the console-conversation-tags store (Conversation-Tags
              domain, MongoDB-elimination EPIC) — the RLS-isolated store backing LibreChat''s conversationTag records. Own-tags-only:
              sharing/marketplace is out of scope. Grants no access to another user''s conversation-tags or to any other memory
              layer.'
            memory:conversation_tags:write: Create/update/delete your own conversation-tags in the console-conversation-tags
...
info:
  description: 'EU-AI-Act-aligned audit trail for LLM workloads. OpenAI-compatible ``/v1/chat/completions`` proxy over a five-layer
    memory model (ADR-062): four per-user-isolated layers — episodic, procedural, conversational, semantic — plus Layer 5,
    the Shared Corpus (org-wide knowledge, shared-read, operator/curator-tier write). Keycloak-delegated identity, Postgres
    RLS per-user isolation, opt-in Redis-Streams async persistence (ADR-046), and a first-class audit event on every memory
    read/list/write/delete (ADR-062 §5 — ``event_class=memory_access`` on ``GET /interactions``).


    Recall (the LLM reading its own memory) happens through the ``recall_*`` tools inside the ``/v1/chat/completions`` tool-call
    loop (ADR-025) and through ``POST /context`` (the legacy ``inject``-mode aggregator) — NOT through the ``memory`` tag''s
    backoffice CRUD endpoints, which are the operator/audit surface for managing memory content directly.


    See ``docs/architecture/`` for C4 + sequence diagrams; the ADR set (014–062) is the source of truth for design decisions.'
  title: AuditTrace-AI
openapi: 3.1.0
paths:
...
  /console/prompts:
    get:
      description: 'List the CALLER''s OWN (non-deleted) prompt groups, newest-first,

        cursor-paginated (summary shape, no versions per row). Never

        returns another user''s group — RLS plus the service''s own explicit

        ``user_sub`` filter.'
      operationId: list_groups_console_prompts_get
      parameters:
      - description: Opaque pagination cursor.
        in: query
        name: cursor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Opaque pagination cursor.
          title: Cursor
      - in: query
        name: limit
        required: false
        schema:
          default: 25
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsolePromptGroupListResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - audittrace-oauth2:
        - memory:prompts:read-own
      summary: List Groups
      tags:
      - console-prompts
    post:
      description: 'Create or update the CALLER''s OWN prompt group (upsert by

        ``group_id``). ``user_sub`` is stamped from ``user`` — the body

        carries no such field to override it with.'
      operationId: upsert_group_console_prompts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsolePromptGroupUpsertRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsolePromptGroupItem'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - audittrace-oauth2:
        - memory:prompts:write
      summary: Upsert Group
      tags:
      - console-prompts
  /console/prompts/{group_id}:
    delete:
      description: 'Soft-delete the CALLER''s OWN prompt group AND its versions. 404

        if not found/not owned/already deleted.'
      operationId: delete_group_console_prompts__group_id__delete
      parameters:
      - in: path
        name: group_id
        required: true
        schema:
          title: Group Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - audittrace-oauth2:
        - memory:prompts:write
      summary: Delete Group
      tags:
      - console-prompts
    get:
      description: 'Fetch the CALLER''s OWN prompt group INCLUDING its full version

        history. 404 if it doesn''t exist, is deleted, or belongs to

        another user — never 403 (would leak existence to a caller probing

        foreign group ids).'
      operationId: get_group_console_prompts__group_id__get
      parameters:
      - in: path
        name: group_id
        required: true
        schema:
          title: Group Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsolePromptGroupWithVersionsItem'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - audittrace-oauth2:
        - memory:prompts:read-own
      summary: Get Group
      tags:
      - console-prompts
  /console/prompts/{group_id}/production:
    patch:
      description: 'Mark ``body.prompt_id`` as the production version of the

        CALLER''s OWN ``group_id``. 404 if the group doesn''t exist/isn''t

        owned, OR the named version doesn''t exist/isn''t owned/doesn''t

        belong to this group — a hostile caller can never point

        ``production_prompt_id`` at another user''s version or at a

        version from a different group.'
      operationId: set_production_console_prompts__group_id__production_patch
      parameters:
      - in: path
        name: group_id
        required: true
        schema:
          title: Group Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsolePromptSetProductionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsolePromptGroupWithVersionsItem'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - audittrace-oauth2:
        - memory:prompts:write
      summary: Set Production
      tags:
      - console-prompts
  /console/prompts/{group_id}/versions:
    post:
      description: 'Create or update the CALLER''s OWN prompt version (upsert by

        ``prompt_id``), attached to the CALLER''s OWN ``group_id``. 404 if

        the group doesn''t exist/isn''t owned by the caller — a hostile

        caller can never attach a version to another user''s group by

        guessing its ``group_id``.'
      operationId: upsert_version_console_prompts__group_id__versions_post
      parameters:
      - in: path
        name: group_id
        required: true
        schema:
          title: Group Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsolePromptVersionUpsertRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsolePromptVersionItem'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - audittrace-oauth2:
        - memory:prompts:write
      summary: Upsert Version
      tags:
      - console-prompts
  /console/tool-favorites:
    get:
      description: 'List ALL of the CALLER''s OWN (non-deleted) tool-favorites,

        oldest-first. Never returns another user''s favorite — RLS plus the

        service''s own explicit ``user_sub`` filter.'
      operationId: list_tool_favorites_console_tool_favorites_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsoleToolFavoriteListResponse'
          description: Successful Response
      security:
      - audittrace-oauth2:
        - memory:tool_favorites:read-own
      summary: List Tool Favorites
      tags:
      - console-tool-favorites
    post:
      description: 'Add (or idempotently re-affirm) the CALLER''s OWN tool-favorite.

        ``user_sub`` is stamped from ``user`` — the body carries no such

        field to override it with. 409 if the caller already owns

        ``MAX_TOOL_FAVORITES`` active favorites and this pair is a new one.'
      operationId: add_tool_favorite_console_tool_favorites_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsoleToolFavoriteAddRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsoleToolFavoriteItem'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - audittrace-oauth2:
        - memory:tool_favorites:write
      summary: Add Tool Favorite
      tags:
      - console-tool-favorites
  /console/tool-favorites/{item_type}/{item_id}:
    delete:
      description: 'Soft-delete the CALLER''s OWN tool-favorite. 404 if not

        found/not owned/already removed.'
      operationId: remove_tool_favorite_console_tool_favorites__item_type___item_id__delete
      parameters:
      - in: path
        name: item_type
        required: true
        schema:
          title: Item Type
          type: string
      - in: path
        name: item_id
        required: true
        schema:
          title: Item Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - audittrace-oauth2:
        - memory:tool_favorites:write
      summary: Remove Tool Favorite
      tags:
      - console-tool-favorites
...